requirements.txt
pygame>=2.5.0
PyOpenGL>=3.1.7
pillow>=11.2.1